fix sintaxis format

+++ tasks/mpi/filateva_e_metod_gausa/src/ops_mpi.cpp
@@ -5,7 +5,7 @@
 
 bool filateva_e_metod_gausa_mpi::MetodGausa::pre_processing() {
   internal_order_test();
-  if (world.rank() == 0){
+  if (world.rank() == 0) {
     size = taskData->inputs_count[0];
     auto* temp = reinterpret_cast<double*>(taskData->inputs[0]);
     this->matrix.insert(matrix.end(), temp, temp + size * size);
@@ -20,7 +20,7 @@ bool filateva_e_metod_gausa_mpi::MetodGausa::pre_processing() {
 
 bool filateva_e_metod_gausa_mpi::MetodGausa::validation() {
   internal_order_test();
-  if (world.rank() == 0){
+  if (world.rank() == 0) {
     return taskData->inputs_count[0] == taskData->outputs_count[0];
   }
   return true;
@@ -39,21 +39,21 @@ bool filateva_e_metod_gausa_mpi::MetodGausa::run() {
   std::vector<double> lU;
   std::vector<double> lL;
 
-  if (world.rank() == 0){
+  if (world.rank() == 0) {
     L.resize(size * size, 0.0);
     L[0] = 1.0;
     U = matrix;
-  }else{
-    t_strock.resize(size,0);
+  } else {
+    t_strock.resize(size, 0);
   }
 
   for (int i = 1; i < size; ++i) {
     delta = (size - i) / (world.size() - 1);
     ost = (size - i) % (world.size() - 1);
 
-    if (world.rank() == 0){
-      if (delta != 0){
-        for (int proc = 0; proc < world.size() - 1; proc++){
+    if (world.rank() == 0) {
+      if (delta != 0) {
+        for (int proc = 0; proc < world.size() - 1; proc++) {
           world.send(proc + 1, 0, U.data() + proc * size * delta + (ost + i) * size,  delta * size);
         }
       }
@@ -61,61 +61,60 @@ bool filateva_e_metod_gausa_mpi::MetodGausa::run() {
       lU.resize(ost * size, 0.0);
       std::copy(U.begin() + i * size, U.begin() + size * (ost + i), lU.begin());
       t_strock.assign(U.begin() + (i - 1) * size, U.begin() + i * size);
-    }else {
+    } else {
       lL.resize(delta, 0.0);
       lU.resize(size * delta, 0.0);
-      if (delta != 0){
+      if (delta != 0) {
         world.recv(0, 0, lU.data(), delta * size);
       }
     }
 
     boost::mpi::broadcast(world, t_strock.data(), size, 0);
 
-    for (int j = 0; j < lL.size(); ++j){
+    for (long unsigned int j = 0; j < lL.size(); ++j) {
       lL[j] = lU[j * size + i - 1] / t_strock[i - 1];
-      for (int k = i - 1; k < size; k++){
+      for (int k = i - 1; k < size; k++) {
         lU[j * size + k] -= lL[j] * t_strock[k];
       }
     }
   
-    if (world.rank() == 0){
-      if (lU.size() != 0){
-        std::copy(lU.begin(), lU.end(), U.begin() + i*size);
-        for (int j = 0;j < lL.size(); j++){
+    if (world.rank() == 0) {
+      if (lU.size() != 0) {
+        std::copy(lU.begin(), lU.end(), U.begin() + i * size);
+        for (long unsigned int j = 0;j < lL.size(); j++) {
           L[(i + j) * size + i - 1] = lL[j];
         }
       }
 
       std::vector<double> temp1(size * delta);
       std::vector<double> temp2(delta);
-      if (delta != 0){
-        for (int proc = 0; proc < 2 * (world.size() - 1); proc++){
+      if (delta != 0) {
+        for (int proc = 0; proc < 2 * (world.size() - 1); proc++) {
           status = world.probe( boost::mpi::any_source,  boost::mpi::any_tag);
           if (status.tag() == 0){
             world.recv(status.source(), status.tag(), temp1.data(), delta * size);
-            std::copy(temp1.begin(),temp1.end(), U.begin() + (status.source() - 1) * size * delta + (ost + i) * size);
-          } 
-          else if (status.tag() == 1){
+            std::copy(temp1.begin(), temp1.end(), U.begin() + (status.source() - 1) * size * delta + (ost + i) * size);
+          } else if (status.tag() == 1) {
             world.recv(status.source(), status.tag(), temp2.data(), delta);
-            for (int j = 0; j < delta; j++){
-              L[(i + ost + (status.source() - 1)* delta +  j) * size + i - 1] = temp2[j];
+            for (int j = 0; j < delta; j++) {
+              L[(i + ost + (status.source() - 1) * delta + j) * size + i - 1] = temp2[j];
             }
           }
         }
       }
-    }else {
-      if (delta != 0){
+    } else {
+      if (delta != 0) {
         world.send(0, 0, lU.data(), delta * size);
         world.send(0, 1, lL.data(), delta);
       }
     }
 
-    if (world.rank() == 0){
+    if (world.rank() == 0) {
       L[i * size + i] = 1.0;
     }
   }
 
-  if (world.rank() == 0){
+  if (world.rank() == 0) {
     std::vector<double> y(size);
     for (int i = 0; i < size; i++) {
       y[i] = b_vector[i];
@@ -132,14 +131,14 @@ bool filateva_e_metod_gausa_mpi::MetodGausa::run() {
       resh[i] /= U[i * size + i];
     }
   }
-  
+
   return true;
 }
 
 bool filateva_e_metod_gausa_mpi::MetodGausa::post_processing() {
   internal_order_test();
-  if (world.rank() == 0){
-    taskData->outputs.emplace_back(reinterpret_cast<uint8_t *>(resh.data()));
+  if (world.rank() == 0) {
+    taskData->outputs.emplace_back(reinterpret_cast<uint8_t*>(resh.data()));
   }
   return true;
 }
